requirements_minimal.txt
# Dependencias mínimas para funcionamiento básico del chatbot
# Instalar con: pip install -r requirements_minimal.txt

# Framework web
Flask==3.1.0
python-dotenv==1.0.0
requests==2.32.3

# Procesamiento de documentos
PyPDF2==3.0.1
python-docx==1.1.2

# Machine Learning básico
sentence-transformers>=3.0.0
scikit-learn>=1.3.0
numpy>=1.24.0

# Google AI (esencial para el chatbot)
google-generativeai>=0.8.0

# Opcional: descomenta si necesitas funcionalidades específicas
# Pillow>=10.0.0  # Para procesamiento básico de imágenes
# opencv-python>=4.8.0  # Para visión computacional
# transformers>=4.40.0  # Para modelos de IA avanzados
# torch>=2.0.0  # Para deep learning